Move error handling and add error message when elasticsearch data is incorrect
