+++ application/controllers/artifacts_controller.py
@@ -8,10 +8,10 @@ import werkzeug
 from flask import current_app
 from webargs.flaskparser import use_args
 from application.errors import NotFound
-import application.controllers.error_handling.request_parsing # pylint: disable=W0611
+import application.error_handling.request_parsing # pylint: disable=unused-import
+from application.error_handling.es_connection import check_es_connection
 from application.base import respond_with
 from application.models.artifact import Artifact
-from application.controllers.error_handling.es_connection import check_es_connection
 from application.validators import artifacts_validator
 from .application_controller import ApplicationController
